working on water bullet sdf changes

+++ src/main/java/com/amuzil/magus/skill/traits/skilltraits/FloatTrait.java
@@ -31,7 +31,7 @@ public class FloatTrait extends SkillTrait {
         markDirty();
     }
 
-    public double getValue() {
+    public float getValue() {
         return val;
     }
